Add SubcategoryPresenter to render 'part_of' links
This adds a new presenter which decorates instances of Subcategory. The
presenter provides a new method, 'part_of_links', which returns an array
of pre-rendered links ready to be dropped into the 'document footer' 
component.

By default, we will only display a link to the parent topic (eg. 'Part 
of: Oil and gas'), but when 'include_child' is set to true, we prepend 
a link to the sub-topic too.