Add an alternative sub-topic layout
We want to try out a new layout for sub-topic pages, which removes the
breadcrumbs and metadata from the page, instead using the right-hand 
box and the new document footer component to orient users around the 
page.

As we want to compare how users interact with this layout, we're 
intending to only display it on a small subset of sub-topic pages.